IllustsBean 已并入 ceui.loxia.Illust，IllustCaptionExporter / ArtworkV3Fragment 新增代码跟着改用 Illust。

+++ models/src/main/java/ceui/lisa/models/UserDetailResponse.java
@@ -51,6 +51,6 @@ public class UserDetailResponse extends UserHolder implements Serializable, User
 
     @Override
     public int getUserId() {
-        return getUser() == null ? 0 : getUser().getId();
+        return getUser() == null ? 0 : (int) getUser().getId();
     }
 }
